+++ src/baobab_scryfall_api_caller/mappers/__init__.py
@@ -7,5 +7,6 @@ from baobab_scryfall_api_caller.mappers.card_mapper import CardMapper
 from baobab_scryfall_api_caller.mappers.scryfall_error_translator import (
     ScryfallErrorTranslator,
 )
+from baobab_scryfall_api_caller.mappers.set_mapper import SetMapper
 
-__all__ = ["ErrorTranslationContext", "CardMapper", "ScryfallErrorTranslator"]
+__all__ = ["ErrorTranslationContext", "CardMapper", "SetMapper", "ScryfallErrorTranslator"]