+++ src/baobab_scryfall_api_caller/models/sets/__init__.py
@@ -1,0 +1,5 @@
+"""Modeles du domaine Sets."""
+
+from baobab_scryfall_api_caller.models.sets.set import Set
+
+__all__ = ["Set"]